Move ShieldLevelHelper inside the Config class
